:lipstick: Rework the UI
